+++ frontend/src/Author/Details/AuthorDetailsSeriesConnector.js
@@ -4,11 +4,7 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
-import { toggleBooksMonitored } from 'Store/Actions/bookActions';
-import { executeCommand } from 'Store/Actions/commandActions';
 import { setSeriesSort, setSeriesTableOption } from 'Store/Actions/seriesActions';
-import createAuthorSelector from 'Store/Selectors/createAuthorSelector';
-import createCommandsSelector from 'Store/Selectors/createCommandsSelector';
 import createDimensionsSelector from 'Store/Selectors/createDimensionsSelector';
 import createUISettingsSelector from 'Store/Selectors/createUISettingsSelector';
 import AuthorDetailsSeries from './AuthorDetailsSeries';
@@ -17,12 +13,10 @@ function createMapStateToProps() {
   return createSelector(
     (state, { seriesId }) => seriesId,
     (state) => state.books,
-    createAuthorSelector(),
     (state) => state.series,
-    createCommandsSelector(),
     createDimensionsSelector(),
     createUISettingsSelector(),
-    (seriesId, books, author, series, commands, dimensions, uiSettings) => {
+    (seriesId, books, series, dimensions, uiSettings) => {
 
       const currentSeries = _.find(series.items, { id: seriesId });
 
@@ -59,7 +53,6 @@ function createMapStateToProps() {
         columns: series.columns,
         sortKey: series.sortKey,
         sortDirection: series.sortDirection,
-        authorMonitored: author.monitored,
         isSmallScreen: dimensions.isSmallScreen,
         uiSettings
       };
@@ -68,10 +61,8 @@ function createMapStateToProps() {
 }
 
 const mapDispatchToProps = {
-  toggleBooksMonitored,
   setSeriesTableOption,
-  dispatchSetSeriesSort: setSeriesSort,
-  executeCommand
+  dispatchSetSeriesSort: setSeriesSort
 };
 
 class AuthorDetailsSeasonConnector extends Component {
@@ -87,13 +78,6 @@ class AuthorDetailsSeasonConnector extends Component {
     this.props.dispatchSetSeriesSort({ sortKey });
   };
 
-  onMonitorBookPress = (bookIds, monitored) => {
-    this.props.toggleBooksMonitored({
-      bookIds,
-      monitored
-    });
-  };
-
   //
   // Render
 
@@ -103,7 +87,6 @@ class AuthorDetailsSeasonConnector extends Component {
         {...this.props}
         onSortPress={this.onSortPress}
         onTableOptionChange={this.onTableOptionChange}
-        onMonitorBookPress={this.onMonitorBookPress}
       />
     );
   }
@@ -111,10 +94,8 @@ class AuthorDetailsSeasonConnector extends Component {
 
 AuthorDetailsSeasonConnector.propTypes = {
   authorId: PropTypes.number.isRequired,
-  toggleBooksMonitored: PropTypes.func.isRequired,
   setSeriesTableOption: PropTypes.func.isRequired,
-  dispatchSetSeriesSort: PropTypes.func.isRequired,
-  executeCommand: PropTypes.func.isRequired
+  dispatchSetSeriesSort: PropTypes.func.isRequired
 };
 
 export default connect(createMapStateToProps, mapDispatchToProps)(AuthorDetailsSeasonConnector);