+++ frontend/src/Bookshelf/BookshelfRowConnector.js
@@ -3,8 +3,6 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
-import { toggleAuthorMonitored } from 'Store/Actions/authorActions';
-import { toggleBooksMonitored } from 'Store/Actions/bookActions';
 import createAuthorSelector from 'Store/Selectors/createAuthorSelector';
 import BookshelfRow from './BookshelfRow';
 
@@ -31,9 +29,7 @@ function createMapStateToProps() {
         ...author,
         authorId: author.id,
         authorName: author.authorName,
-        monitored: author.monitored,
         status: author.status,
-        isSaving: author.isSaving,
         books: sortedBooks
       };
     }
@@ -41,34 +37,9 @@ function createMapStateToProps() {
 }
 
 const mapDispatchToProps = {
-  toggleAuthorMonitored,
-  toggleBooksMonitored
 };
 
 class BookshelfRowConnector extends Component {
-
-  //
-  // Listeners
-
-  onAuthorMonitoredPress = () => {
-    const {
-      authorId,
-      monitored
-    } = this.props;
-
-    this.props.toggleAuthorMonitored({
-      authorId,
-      monitored: !monitored
-    });
-  };
-
-  onBookMonitoredPress = (bookId, monitored) => {
-    const bookIds = [bookId];
-    this.props.toggleBooksMonitored({
-      bookIds,
-      monitored
-    });
-  };
 
   //
   // Render
@@ -77,18 +48,13 @@ class BookshelfRowConnector extends Component {
     return (
       <BookshelfRow
         {...this.props}
-        onAuthorMonitoredPress={this.onAuthorMonitoredPress}
-        onBookMonitoredPress={this.onBookMonitoredPress}
       />
     );
   }
 }
 
 BookshelfRowConnector.propTypes = {
-  authorId: PropTypes.number.isRequired,
-  monitored: PropTypes.bool.isRequired,
-  toggleAuthorMonitored: PropTypes.func.isRequired,
-  toggleBooksMonitored: PropTypes.func.isRequired
+  authorId: PropTypes.number.isRequired
 };
 
 export default connect(createMapStateToProps, mapDispatchToProps)(BookshelfRowConnector);